requirements.txt
flask
flask-cors
gunicorn
sentence-transformers
faiss-cpu
pandas
numpy
deep-translator